requirements.txt
pytest==7.2.0
selenium==4.6.0
allure-pytest==2.12.0
requests~=2.28.2
